Renaming the service folder into serviceS
